Moved filterbar to ui folder
